feat: cover obs sdk by 3.23.9
