+++ openstack/obs/client_part.go
@@ -115,8 +115,8 @@ func (obsClient ObsClient) UploadPart(_input *UploadPartInput, extensions ...ext
 	output = &UploadPartOutput{}
 	var repeatable bool
 	if input.Body != nil {
-		if _, ok := input.Body.(*strings.Reader); !ok {
-			repeatable = false
+		if _, ok := input.Body.(*strings.Reader); ok {
+			repeatable = true
 		}
 		if _, ok := input.Body.(*readerWrapper); !ok && input.PartSize > 0 {
 			input.Body = &readerWrapper{reader: input.Body, totalCount: input.PartSize}
