+++ openstack/obs/conf.go
@@ -21,10 +21,14 @@ import (
 	"net"
 	"net/http"
 	"net/url"
+	"os"
 	"sort"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
+
+	"golang.org/x/net/http/httpproxy"
 )
 
 type urlHolder struct {
@@ -49,6 +53,8 @@ type config struct {
 	finalTimeout      int
 	maxRetryCount     int
 	proxyURL          string
+	noProxyURL        string
+	proxyFromEnv      bool
 	maxConnsPerHost   int
 	pemCerts          []byte
 	transport         *http.Transport
@@ -58,6 +64,10 @@ type config struct {
 	maxRedirectCount  int
 	userAgent         string
 	enableCompression bool
+	progressListener  ProgressListener
+
+	customProxyOnce      sync.Once
+	customProxyFuncValue func(*url.URL) (*url.URL, error)
 }
 
 func (conf config) String() string {
@@ -106,6 +116,20 @@ func WithHeaderTimeout(headerTimeout int) configurer {
 func WithProxyUrl(proxyURL string) configurer {
 	return func(conf *config) {
 		conf.proxyURL = proxyURL
+	}
+}
+
+// WithNoProxyUrl is a configurer for ObsClient to set HTTP no_proxy.
+func WithNoProxyUrl(noProxyURL string) configurer {
+	return func(conf *config) {
+		conf.noProxyURL = noProxyURL
+	}
+}
+
+// WithProxyFromEnv is a configurer for ObsClient to get proxy from evironment.
+func WithProxyFromEnv(proxyFromEnv bool) configurer {
+	return func(conf *config) {
+		conf.proxyFromEnv = proxyFromEnv
 	}
 }
 
@@ -342,14 +366,9 @@ func (conf *config) getTransport() error {
 			ResponseHeaderTimeout: time.Second * time.Duration(conf.headerTimeout),
 			IdleConnTimeout:       time.Second * time.Duration(conf.idleConnTimeout),
 		}
-
 		if conf.proxyURL != "" {
-			proxyURL, err := url.Parse(conf.proxyURL)
-			if err != nil {
-				return err
-			}
-			conf.transport.Proxy = http.ProxyURL(proxyURL)
-		} else {
+			conf.transport.Proxy = conf.customProxyFromEnvironment
+		} else if conf.proxyFromEnv {
 			conf.transport.Proxy = http.ProxyFromEnvironment
 		}
 
@@ -365,6 +384,24 @@ func (conf *config) getTransport() error {
 	}
 
 	return nil
+}
+
+func (conf *config) customProxyFromEnvironment(req *http.Request) (*url.URL, error) {
+	url, err := conf.customProxyFunc()(req.URL)
+	return url, err
+}
+
+func (conf *config) customProxyFunc() func(*url.URL) (*url.URL, error) {
+	conf.customProxyOnce.Do(func() {
+		customhttpproxy := &httpproxy.Config{
+			HTTPProxy:  conf.proxyURL,
+			HTTPSProxy: conf.proxyURL,
+			NoProxy:    conf.noProxyURL,
+			CGI:        os.Getenv("REQUEST_METHOD") != "",
+		}
+		conf.customProxyFuncValue = customhttpproxy.ProxyFunc()
+	})
+	return conf.customProxyFuncValue
 }
 
 func checkRedirectFunc(req *http.Request, via []*http.Request) error {
@@ -504,4 +541,21 @@ func getQueryURL(key, value string) string {
 	queryURL += key
 	queryURL += value
 	return queryURL
+}
+
+var once sync.Once
+
+func (obsClient ObsClient) GetClientConfigure(extensions []extensionOptions) *config {
+	once.Do(func() {
+		for _, extension := range extensions {
+			if configure, ok := extension.(configurer); ok {
+				configure(obsClient.conf)
+			}
+		}
+	})
+	return obsClient.conf
+}
+
+func (obsClient ObsClient) getProgressListener(extensions []extensionOptions) ProgressListener {
+	return obsClient.GetClientConfigure(extensions).progressListener
 }
