+++ openstack/obs/model_object.go
@@ -9,6 +9,7 @@
 // under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 // CONDITIONS OF ANY KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations under the License.
+
 package obs
 
 import (
@@ -157,6 +158,7 @@ type GetObjectMetadataInput struct {
 // GetObjectMetadataOutput is the result of GetObjectMetadata function
 type GetObjectMetadataOutput struct {
 	BaseModel
+	HttpHeader
 	VersionId               string
 	WebsiteRedirectLocation string
 	Expiration              string
@@ -192,6 +194,7 @@ type GetObjectInput struct {
 	GetObjectMetadataInput
 	IfMatch                    string
 	IfNoneMatch                string
+	AcceptEncoding             string
 	IfUnmodifiedSince          time.Time
 	IfModifiedSince            time.Time
 	RangeStart                 int64
@@ -231,15 +234,14 @@ type ObjectOperationInput struct {
 	Expires                 int64
 	SseHeader               ISseHeader
 	Metadata                map[string]string
+	HttpHeader
 }
 
 // PutObjectBasicInput defines the basic object operation properties
 type PutObjectBasicInput struct {
 	ObjectOperationInput
-	ContentType     string
-	ContentMD5      string
-	ContentLength   int64
-	ContentEncoding string
+	ContentMD5    string
+	ContentLength int64
 }
 
 // PutObjectInput is the input parameter of PutObject function
@@ -271,6 +273,7 @@ type PutObjectOutput struct {
 	StorageClass StorageClassType
 	ETag         string
 	ObjectUrl    string
+	CallbackBody
 }
 
 // CopyObjectInput is the input parameter of CopyObject function
