+++ openstack/obs/model_other.go
@@ -9,6 +9,7 @@
 // under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 // CONDITIONS OF ANY KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations under the License.
+
 package obs
 
 import (
@@ -20,6 +21,7 @@ type CreateSignedUrlInput struct {
 	Method      HttpMethodType
 	Bucket      string
 	Key         string
+	Policy      string
 	SubResource SubResourceType
 	Expires     int
 	Headers     map[string]string
