+++ openstack/obs/model_part.go
@@ -9,6 +9,7 @@
 // under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 // CONDITIONS OF ANY KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations under the License.
+
 package obs
 
 import (
@@ -114,6 +115,7 @@ type CompleteMultipartUploadOutput struct {
 	Key          string     `xml:"Key"`
 	ETag         string     `xml:"ETag"`
 	EncodingType string     `xml:"EncodingType,omitempty"`
+	CallbackBody
 }
 
 // ListPartsInput is the input parameter of ListParts function
