+++ openstack/obs/temporary_createSignedUrl.go
@@ -52,7 +52,7 @@ func (obsClient ObsClient) CreateSignedUrl(input *CreateSignedUrlInput, extensio
 		input.Expires = 300
 	}
 
-	requestURL, err := obsClient.doAuthTemporary(string(input.Method), input.Bucket, input.Key, params, headers, int64(input.Expires))
+	requestURL, err := obsClient.doAuthTemporary(string(input.Method), input.Bucket, input.Key, input.Policy, params, headers, int64(input.Expires))
 	if err != nil {
 		return nil, err
 	}
